Added Breakpoints manager and identify them differently
This is introduce as a first step to support using of loaded breakpoints on reconnecting to a new process.

When the web UI still runs, it is useful to keep previous breakpoints for debugging, but this needs a more reliable approach. We now use the source URI and line number or the coordinates of the source section. That way, we can restablish breakpoints even when ids don’t match.

This does not yet adapt the front end

Signed-off-by: Stefan Marr <git@stefan-marr.de>